requirements.txt
fastapi
uvicorn
openai
python-multipart
pydantic
faster-whisper
google-generativeai
ffmpeg
streamlit 
requests
watchdog